Introduce mock library and fix integration test
Integration tests are failing with the default local_settings.py file
because the PEAM_AUTH_BASE_URL setting is set to a non-valid url. This
value should be overridden in tests, and the only way to achieve that is
to patch the object at test time. To do so, we make use of the standard
'mock' library, which thus needs to be installed.