Access Redis through Sentinel
This allows us to use a more reliable master-slave Redis cluster. The
Redis Sentinel knows the location of the redis master and shares it with
the flask app. If the master goes down, the slave becomes master.

Redis and sentinel are run inside a docker container. The docker-compose
file can be used both in development (in a virtual machine or not) and
in production.