Don't log state-related PEAM errors due to clients
AuthStateMissing are due to a missing session state parameter.

AuthMissingParameter is triggered when the request does not contain a
valid state or redirect_state parameter.

Both these errors are caused by client or 3rd-party behaviour, so we
should not log them.

This should solve https://opbeat.com/labonneboite/lbb-prod/errors/2906/