Catch invalid PEAM return values
Sometimes, PEAM answers with missing values (gender, first name, ...).
In such cases, we should not crash with a KeyError, but catch the error
and signal it to the user.

For now, we also log the error because we want to be warned.

This will suppress https://opbeat.com/labonneboite/lbb-prod/errors/57/