Remove form indirection
Previously, we were redirecting search.recherche to search.entreprises.
This was done to translate url query string parameters from a long to a
short form (e.g: "location" -> "l"). This was causing a lot of
complexity so we removed it.

As a consequence, distance=5km is no longer accepted in the frontend. We
must fix this because we receive a lot of requests from PE.fr for which
distance=5 km.