Allow only some file extensions for Tilkee upload
The html input element for file selection in the Tilkee upload form
already contains an "accept" attribute that is supposed to limit the
file types uploaded to Tilkee. However, this attribute is not supported
by Edge and Safari. Thus, we manually enforce file types on the client
side and the server side.