Transparent single sign-on
Users are transparently signed-on to their Pole Emploi account through
Pe-Connect. This is achieved by a invisible iframe that loads the
authentication page. The user account header is then refreshed
asynchronously if the user is correctly connected.

Note that some features will not be activated on the same page where
login was achieved. For instances, the "add to favorites" buttons will
be immediately visible.