Add users connected on pole-emploi.fr to GA
The 'idutkes' cookie given to users after they are connected to their
account on pole-emploi.fr is on the domain '.pole-emploi.fr'. Thus, it
is possible to read this cookie on labonneboite.pole-emploi.fr. By
checking the presence of this cookie, we will be able to detect users
that are connected to their pe.fr account, but who are not pe-connected
yet.