On office_admin_update creation, check if company already have one
 Sometimes, recruiter send two emails at the same time. Both conatins a create office_admin_update link.
But, when the first office_admin_update is created, the suggestion in the second mail became wrong.
And, on saving, we got a conflict with the recently created office_admin_update.
To avoid that, we redirect to the edit page if, afterall, an office_admin_update already exists for the company siret.