Display Openstreetmap attribution on maps
It is surprisingly difficult to override the mapbox-gl.js attribution.
For that, we need to subclass the AttributionControl class. If we don't,
then there is an attribution to Openmaptiles that is added to the map,
and we don't want this attribution because we do not rely on their
servers.

We took the opportunity to upgrade mapbox-gl.js from 0.43.0 to 0.53.1
(latest version at this date) by adding the vendor library to the repo.
Thus, we no longer pull the js library from the remote server. This
should help running LBB locally with no internet access (although tiles
will not work, of course).