Attempt to fix IGN timeouts on duration endpoint
The IGN API timeouts whenever we try to load durations for the 15
companies on each result page. When this occurs, no duration is
displayed on the result page. We address this by splitting the calls in
bundles of 5 companies: IGN should reply faster, and if it doesn't then
at least some of the companies should have a proper duration.

The drawback is that this causes 3 /durations request instead of just
one, both for us and the IGN.