Refacto GeocodeUnit
It does not make sense anymore to have a class, for this simple function.
We removed it, turned it into a simple function.
It makes the debug easier, and the code much more readable.
We also fixed some little bugs, according to the geocoding which ran for
the june importer.