Fix threshold score minimum for rome which are metiers tension
If a rome is a "metier en tension", we want that a company which offers
this kind of job, appears more on LBB. For this, we will low down threshold if a metier is
strongly in tension.
We use a file which has been built by Julie,
and which matches for each rome a tension percentage.