Fix isochrone precompute SPOF
Whenever redis master was not available, precompute became impossible.
This was causing 500 errors on /entreprises search and partial
unavailability of the service. Here, we fix the 500 error and log an
error message instead. This does not solve the root cause of the redis
connection error, but at least we can deal with it.