Add history table to keep track of importer jobs execution
For the deployment of importer on k8s stack, the jobs need to be sure
that the previous one has been executed.
There won't be any jenkins for orchestration of jobs.
And, there is also antoher reason.
With this, we can store exceptions that occure in importer in the database !